+++ src/utils/index.ts
@@ -1,4 +1,6 @@
+export * from './animate-waypoints';
 export * from './clamp';
+export * from './haversine-distance';
 export * from './lng-lat-alt-to-vector3';
 export * from './make-perspective-matrix';
 export * from './projected-units-per-meter';